+++ pkgs/test/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.14.4-dev
+
+* Update browser/node bootstrapping logic to ensure the bootstrap library has
+  the same language version as the test.
+
 ## 1.14.3
 
 * Fix an issue where coverage tests could not run in Chrome headless.