+++ pkgs/test_api/CHANGELOG.md
@@ -5,6 +5,8 @@
     behavior of the code regarding to handling of nulls.
   * **Breaking Change**: `GroupEntry.name` is no longer nullable, the root
     group now has the empty string as its name.
+* Add `languageVersionComment` on the `MetaData` class. This should only be
+  presen for test suites.
 * Deprecate `LiveTestController.liveTest`, the `LiveTestController` instance now
   implements `LiveTest` and can be used directly.